% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.r
\name{dom}
\alias{dom}
\title{Real Value Degree of Membership to a Triangular Fuzzy Number}
\usage{
dom(x, TFN)
}
\arguments{
\item{x}{a numeric vector.}

\item{TFN}{a numeric vector of length 3.}
}
\value{
Returns a numeric in interval \code{[0,1]}.
}
\description{
Calculates the degree of membership of a real number to a 
triangular fuzzy number. The fuzzy number is defined by its 
central value and the left and right spreads.
}
\examples{
x <- seq(from = 0, to = 2, length.out = 10)
A <- c(1, 1, 1)
dom(x, A)
}
