% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fuzzylm.R
\name{print.fuzzylm}
\alias{print.fuzzylm}
\title{Prints Fuzzy Linear Regression Result}
\usage{
\method{print}{fuzzylm}(x, ...)
}
\arguments{
\item{x}{a \code{fuzzylm} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the call and coefficients from the \code{fuzzylm} object.
}
\examples{
x <- rep(1:3, each = 5)
y <- c(rnorm(5, 1), rnorm(5, 2), rnorm(5, 3))
dat <- data.frame(x = x, y = y)
f <- fuzzylm(y ~ x, dat)
f
}
\keyword{fuzzy}
