% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumFuzzy.R
\name{sumFuzzy}
\alias{sumFuzzy}
\title{Sum of Two Triangular Fuzzy Numbers}
\usage{
sumFuzzy(x, y)
}
\arguments{
\item{x}{a numeric vector of length three, specifying a triangular fuzzy number as its 
central value, left and right spread.}

\item{y}{a numeric vector of length three, specifying a triangular fuzzy number as its 
central value, left and right spread.}
}
\value{
Returns a numeric vector, representing a triangular fuzzy number.
}
\description{
Calculates a sum of two triangular fuzzy numbers defined as a 
central value, left and right spread.
}
\examples{
  x <- c(1, 0.1, 0.2)
  y <- c(2, 0.2, 0.2)
  sumFuzzy(x = x, y = y)
}
