% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fuzzylm.R
\name{summary.fuzzylm}
\alias{summary.fuzzylm}
\title{Summarizes Fuzzy Linear Regression}
\usage{
\method{summary}{fuzzylm}(object, ...)
}
\arguments{
\item{object}{a \code{fuzzylm} object.}

\item{...}{additional parameters passed to and from other methods.}
}
\value{
Returns a list with models for the central tendency and spreads from the fuzzy
  linear regression, total error of fit of the model and a goodness-of-fit measure.
}
\description{
Calculates the summary from the \code{fuzzylm} object.
}
\examples{
data(fuzzydat)
f <- fuzzylm(y ~ x, fuzzydat$lee)
sum.f <- summary(f)
sum.f
}
\seealso{
\code{\link{TEF}}, \code{\link{GOF}}
}
\keyword{fuzzy}
